refactor:(api) refactor video and subscription service and validation logic
* Also added swagger docs for download video route

+++ src/components/tweet/tweet.controller.js
@@ -1,6 +1,9 @@
+import { ApiError } from "../../utils/apiErrors.js";
 import { ApiResponse } from "../../utils/apiResponse.js";
 import { asyncHandler } from "../../utils/handlers.js";
 import * as tweetService from "./tweet.service.js";
+import debug from "debug";
+const log = debug("app:tweet:controller:log");
 
 const createTweet = asyncHandler(async (req, res) => {
 	const { content, title } = req.body;
@@ -13,6 +16,8 @@ const createTweet = asyncHandler(async (req, res) => {
 		contentImage,
 	);
 
+   log("Tweet created successfully", tweet);
+
 	return res
 		.status(201)
 		.json(new ApiResponse(201, tweet, "Successfully create tweet"));
@@ -22,22 +27,39 @@ const updateTweet = asyncHandler(async (req, res) => {
 	const { tweetId } = req.params;
 	const { content, title } = req.body;
 
-	const tweet = await tweetService.updateTweet(content, title, tweetId);
+	const tweet = await tweetService.findTweetById(tweetId);
+	if (!tweet) throw new ApiError(404, "Tweet not found");
+
+	const isOwner = await tweetService.isTweetOwner(tweetId, req.user._id);
+	if (!isOwner) throw new ApiError(402, "Unauthorized to perform this task");
+
+
+	const updatedTweet = await tweetService.updateTweet(content, title, tweetId);
 	return res
 		.status(200)
-		.json(new ApiResponse(200, tweet, "Successfully updated tweet"));
+		.json(new ApiResponse(200, updatedTweet, "Successfully updated tweet"));
 });
 
 const deleteTweet = asyncHandler(async (req, res) => {
 	const { tweetId } = req.params;
+
+	const tweet = await tweetService.findTweetById(tweetId);
+	if (!tweet) throw new ApiError(404, "Tweet not found");
+
+	const isOwner = await tweetService.isTweetOwner(tweetId, req.user._id);
+	if (!isOwner) throw new ApiError(402, "Unauthorized to perform this task");
+
 	await tweetService.deleteTweet(tweetId);
 	return res.status(204).end();
 });
 
 const getUserTweets = asyncHandler(async (req, res) => {
 	const { userId } = req.params;
-	const user = userId || req.user.Id;
-	const tweets = await tweetService.getUserTweets(user);
+
+   const user = await tweetService.findUserById(userId);
+   if (!user) throw new ApiError(404, "User not found");
+
+	const tweets = await tweetService.getUserTweets(user._id);
 
 	return res
 		.status(200)
